chore: prepare for F-Droid release
#1035
