+++ .github/workflows/test-e2e.yml
@@ -59,7 +59,7 @@ jobs:
           disable-animations: true
           sdcard-path-or-size: 100M
           script: |
-            ./gradlew installDebug
+            ./gradlew installFDroidDebug
             adb shell settings put system show_touches 1
             $HOME/.maestro/bin/maestro test .maestro
