+++ .github/workflows/update-compose-compiler-report.yml
@@ -45,7 +45,7 @@ jobs:
         shell: bash
 
       - name: Assemble apk with compose compiler reports enabled
-        run: ./gradlew assembleRelease -Pmusicsearch.enableComposeCompilerReports=true --rerun-tasks
+        run: ./gradlew assembleFDroidRelease -Pmusicsearch.enableComposeCompilerReports=true --rerun-tasks
 
       - name: Generate HTML from compose reports
         run: find . -type d -path '*/build/compose_metrics' | xargs -I {} composeReport2Html -app MusicSearch -i {} -o {}
