+++ fastlane/metadata/android/en-US/full_description.txt 
@@ -1,0 +1,11 @@
+Features:
+- Search MusicBrainz's massive database for any information related to your favorite artist or song
+- Offline-first; all data is cached on device after loading each page/tab
+- See every page you've visited in the history screen, and quickly get back to them
+- Almost every tab allows you to filter its content instantaneously
+- Save anything to a collection
+- Login using your MusicBrainz account to add to your existing collections
+- Listening on Spotify? Enable Device Broadcast Status to search the artist or song from the app
+- Have a Pixel phone? Enable notification listener to record Now Playing history
+- Dark theme
+- Material You theme
