+++ fastlane/metadata/android/en-US/images/featureGraphic.png
